requirements.txt
streamlit
python-dotenv
youtube-transcript-api
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-google-genai
faiss-cpu
sentence-transformers